+++ hwtest/tests/pci_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from hwtest.tests.helpers import run_command
+from hwtest.tests.shell_utils import run_command
 
 # fmt: off
 params = {
