+++ hwtest/tests/vl805_test.py
@@ -1,4 +1,4 @@
-from hwtest.tests.helpers import run_command
+from hwtest.tests.shell_utils import run_command
 
 
 def test_vl805_fw():
